Cleanup in id module

+++ id/src/main/java/com/ledmington/cpu/x86/RelativeOffset.java
@@ -9,11 +9,11 @@ public final class RelativeOffset implements Operand {
     private final long value;
 
     public static RelativeOffset of8(final byte x) {
-        return new RelativeOffset((long) x);
+        return new RelativeOffset(x);
     }
 
     public static RelativeOffset of32(final int x) {
-        return new RelativeOffset((long) x);
+        return new RelativeOffset(x);
     }
 
     private RelativeOffset(final long value) {
